requirements.txt
azure-functions
tzdata
reportlab